Перекинул выбор расписания из Activity в Fragment, начал химичить с макетами
