+++ Rasp3/app/src/main/AndroidManifest.xml
@@ -43,7 +43,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
-            android:name=".ui.schedule.SelectScheduleFragment"
+            android:name=".ui.schedule.ScheduleFragment"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
