+++ Rasp3/app/src/main/res/menu/main.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/action_select_schedule"
+    <!--<item android:id="@+id/action_select_schedule"
         android:title="@string/action_select_schedule"
         android:orderInCategory="100"
-        app:showAsAction="never" />
+        app:showAsAction="never" />-->
     <item android:id="@+id/action_create_lesson"
         android:title="Добавить пару"
         android:orderInCategory="200"
