+++ Rasp3/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/nav_schedule"
-        android:name="sey.a.rasp3.ui.schedule.ScheduleFragment"
+        android:name="sey.a.rasp3.ui.lesson.LessonFragment"
         android:label="@string/menu_schedule"
         tools:layout="@layout/fragment_schedule" />
 
@@ -40,4 +40,9 @@
         android:name="sey.a.rasp3.ui.send.SendFragment"
         android:label="@string/menu_send"
         tools:layout="@layout/fragment_send" />
+
+    <fragment
+        android:id="@+id/nav_schedule_select"
+        android:name="sey.a.rasp3.ui.schedule.ScheduleFragment"
+        android:label="Выбрать расписание" />
 </navigation>
